# fte/specs/course-companion/contracts/openapi-v1.yaml
openapi: "3.0.3"
info:
  title: Course Companion FTE — Phase 1 API
  version: "1.0.0"
  description: >
    Zero-LLM backend for Course Companion FTE.
    All endpoints return structured data only.
    No LLM calls are made in this phase.
servers:
  - url: http://localhost:8000
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Question:
      type: object
      properties:
        id: { type: string }
        text: { type: string }
        type: { type: string, enum: [mcq, truefalse, fill] }
        options: { type: array, items: { type: string }, nullable: true }
    Quiz:
      type: object
      properties:
        id: { type: string }
        chapter_id: { type: string }
        title: { type: string }
        questions:
          type: array
          items: { $ref: "#/components/schemas/Question" }
    QuizResult:
      type: object
      properties:
        score: { type: number, description: "0.0 to 100.0" }
        results:
          type: array
          items:
            type: object
            properties:
              question_id: { type: string }
              correct: { type: boolean }
              student_answer: { type: string }
...
paths:
...
  /api/v1/quizzes/{quiz_id}:
    get:
      summary: Get quiz questions (NO answer key)
      security:
        - BearerAuth: []
      parameters:
        - name: quiz_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Quiz with questions only — answer key never included
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Quiz" }
        "401":
          description: Unauthorized
  /api/v1/quizzes/{quiz_id}/submit:
    post:
      summary: Submit quiz answers for server-side grading
      security:
        - BearerAuth: []
      parameters:
        - name: quiz_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: object
                  additionalProperties: { type: string }
                  description: "{ question_id: student_answer }"
      responses:
        "200":
          description: Grading result — answer key never in response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuizResult" }
        "401":
          description: Unauthorized
        "409":
          description: Quiz already submitted
...